Fixing False jobcard overlap error
added an extra layer to find set of sequential job cards so that an instantaneous overlapping can be more practical. 
For example : 
We are finding overlapping for time 2023-11-22 10:20:00 to 2023-11-22 11:20:00.
Current algorithm gives the following job card time slots which are overlapping  
1. 2023-11-22 10:20:00 to 2023-11-22 10:40:00
2. 2023-11-22 10:40:00 to 2023-11-22 10:45:00.
3.2023-11-22 10:50:00 to 2023-11-22 11:00:00.
4.2023-11-22 11:00:00 to 2023-11-22 11:10:00.
and system throws an error for overlapping job cards. (Here production capacity is 3)
So if we look closely all above job cards are sequential so at any moment of time only 1 out of 4 is being operated.
Which means we have 2 more capacity left and system is throwing and error of overlapping which is false.

This extra layer will help to identify such sequential Job cards and group them into such a way that we have their counting as 1. all such groups are added and then  we will check with production capacity.